.history/requirements_20251217121132.txt
fastapi
uvicorn
pandas
requests
numpy
faiss-cpu
torch
sentence-transformers==2.6.1
transformers==4.38.2
huggingface-hub==0.24.7